Optimize container performance with caching and reflection improvements
